@(dynamic)variable shouldn't break fontlock
Currently `foo/some-var` will fontlock `foo`, but if this item is
something that can be dereffed and you type `@foo/some-var`, `foo` is no
longer font-locked.

Similarly: `*dynamic-var*` is font locked, but `@` breaks the regex.

In this change the `@` sign isn't font locked, but the var that follows
it is.

Conflicts:
	clojure-mode.el

Add tests for dynamic-var and ns/refer fontlock