Add fixed indent rules
1 problem occurs when indenting in the middle of a populated list like

(a b | c)

where | is the cursor.

It should indent to

(a b
  |
  c)

but instead we get

(a b
 |
  c)

Because the "(list_lit . (sym_lit) _* @indent)" query is not matching,
apparently. I am unsure why it doesn't match, testing manually indicates
that it should, but I'm not sure what node is being used exactly.