Fix indentation of function calls when treesit node is nil
Previously indenting 2 spaces for function calls only worked when we
were indenting where some node was after the point. Indentation fell
back to picking up the "parent 'list_lit' then indent 1 rule" because
node can sometime be nil and there was nothing to capture in the query.
Queries cannot bind nil to a captured var, simply not possible.

We don't really care what the node at point is, instead we just want to
know the parent of wherever the point is. A function is used to make
the match instead of a query. The function can ignore node and focus
only on the parent node.

This will most likely be an important strategy in semantic indentation.