Highlight regular expressions with tree-sitter-regex grammar
This grammar is bundled in nixos by default and seems good enough for
java regular expressions. It is also maintained under the tree-sitter
github org so is "official".

In order to property identify the #" and closing " characters we have to
parse them with the clojure grammar (in case the regex grammar is not
available) and again with the regex grammar as part of the actual
pattern. This could be avoided if either the clojure grammar captured a
node for the inner contents of the regex literal, or the
treesit-range-settings supported some kind of offest argument like the
neovim tree-sitter mechanisms do.

Should address issue #11

I think that multiple parsers per buffer may be too buggy to use right
now. There are situations where no regex will be present on in a buffer,
but the entire buffer will be highlighted as a regular expression. This
functionality probably needs upstream work in Emacs before we can merge
this into the main branch of clojure-ts-mode