Slightly speed-up clojure-ts-align and fix a minor issue
Before we were trying to align multiple nodes starting from the top-level node
and moving to the most deeply nested node. This could produce misaligned forms
if nested nodes have extra spaces that has to be cleaned up. Now we start from
the most deeply nested node and gradually move to the top of the tree.